refactor: enforce the architectural rule 'never mix stateful objects and stateless helper classes containing static methods'
