+++ src/main/java/spoon/support/visitor/java/reflect/RtMethod.java
@@ -17,14 +17,9 @@
 package spoon.support.visitor.java.reflect;
 
 import java.lang.annotation.Annotation;
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.lang.reflect.TypeVariable;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
-
-import spoon.SpoonException;
 
 public class RtMethod {
 	private Class<?> clazz;
@@ -126,62 +121,7 @@ public class RtMethod {
 		return getDeclaringClass().getName().hashCode() ^ getName().hashCode();
 	}
 
-	public static RtMethod create(Method method) {
-		return new RtMethod(method.getDeclaringClass(), method.getName(), method.getReturnType(),
-				method.getTypeParameters(), method.getParameterTypes(), method.getExceptionTypes(), method.getModifiers(),
-				method.getDeclaredAnnotations(), method.getParameterAnnotations(), method.isVarArgs(),
-				//spoon is compatible with Java 7, so compilation fails here
-				//method.isDefault());
-				_java8_isDefault(method));
-	}
-
-	private static Method _method_isDefault;
-	static {
-		try {
-			_method_isDefault = Method.class.getMethod("isDefault");
-		} catch (NoSuchMethodException | SecurityException e) {
-			//spoon is running with java 7 JDK
-			_method_isDefault = null;
-		}
-	}
-
-	private static boolean _java8_isDefault(Method method) {
-		if (_method_isDefault == null) {
-			//spoon is running with java 7 JDK, all methods are not default, because java 7 does not have default methods
-			return false;
-		}
-		try {
-			return (Boolean) _method_isDefault.invoke(method);
-		} catch (IllegalAccessException | IllegalArgumentException e) {
-			throw new SpoonException("Calling of Java8 Method#isDefault() failed", e);
-		} catch (InvocationTargetException e) {
-			throw new SpoonException("Calling of Java8 Method#isDefault() failed", e.getTargetException());
-		}
-	}
-
-	public static <T> RtMethod[] methodsOf(Class<T> clazz) {
-		final RtMethod[] methods = new RtMethod[clazz.getDeclaredMethods().length];
-		int i = 0;
-		for (Method method : clazz.getDeclaredMethods()) {
-			methods[i++] = create(method);
-		}
-		return methods;
-	}
-
-	public static <T> RtMethod[] sameMethodsWithDifferentTypeOf(Class<T> superClass, List<RtMethod> comparedMethods) {
-		final List<RtMethod> methods = new ArrayList<>();
-		for (Method method : superClass.getDeclaredMethods()) {
-			final RtMethod rtMethod = create(method);
-			for (RtMethod potential : comparedMethods) {
-				if (potential.isLightEquals(rtMethod) && !rtMethod.returnType.equals(potential.returnType)) {
-					methods.add(rtMethod);
-				}
-			}
-		}
-		return methods.toArray(new RtMethod[methods.size()]);
-	}
-
-	private boolean isLightEquals(RtMethod rtMethod) {
+	public boolean isLightEquals(RtMethod rtMethod) {
 		if (this == rtMethod) {
 			return true;
 		}
