+++ src/main/java/spoon/testing/utils/ProcessorUtils.java
@@ -16,11 +16,16 @@
  */
 package spoon.testing.utils;
 
+import org.apache.log4j.Level;
 import spoon.Launcher;
 import spoon.processing.Processor;
+import spoon.processing.ProcessorProperties;
+import spoon.processing.Property;
 import spoon.reflect.factory.Factory;
 import spoon.support.compiler.jdt.JDTBasedSpoonCompiler;
+import spoon.support.util.RtHelper;
 
+import java.lang.reflect.Field;
 import java.util.Collection;
 
 public final class ProcessorUtils {
@@ -31,5 +36,27 @@ public final class ProcessorUtils {
 	public static void process(Factory factory, Collection<Processor<?>> processors) {
 		final JDTBasedSpoonCompiler compiler = (JDTBasedSpoonCompiler) new Launcher().createCompiler(factory);
 		compiler.process(processors);
+	}
+
+	public static void initProperties(Processor<?> p, ProcessorProperties properties) {
+		if (properties != null) {
+			for (Field f : RtHelper.getAllFields(p.getClass())) {
+				if (f.isAnnotationPresent(Property.class)) {
+					Object obj = properties.get(f.getType(), f.getName());
+					if (obj != null) {
+						f.setAccessible(true);
+						try {
+							f.set(p, obj);
+						} catch (Exception e) {
+							Launcher.LOGGER.error(e.getMessage(), e);
+						}
+					} else {
+						p.getFactory().getEnvironment().report(p, Level.WARN,
+								"No value found for property '" + f.getName() + "' in processor " + p.getClass()
+										.getName());
+					}
+				}
+			}
+		}
 	}
 }
